blog_backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-decouple==3.8
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.0